# api-docs/api-docs_v3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "core-rest-services",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    },
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).\n\nCore REST Services"
  },
  "paths": {
...
    "/filesystem": {
      "get": {
        "summary": "/filesystem",
        "parameters": [
          {
            "name": "file",
            "in": "query",
            "example": "./admin-audit.csv",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "delete",
            "in": "query",
            "example": "false",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "example": "Time,Source,User,Resource,Name/ID,Action,Field,Old value,New value\nThu Sep 08 2022 19:05:23 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,Brent Creech,update,accessLevel,\"{\"\"accessLevel\"\":\"\"administrator\"\"}\",\"{\"\"accessLevel\"\":\"\"manage\"\"}\"\nThu Sep 08 2022 19:05:14 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,Brent Creech,update,disabled,true,\nThu Sep 08 2022 19:03:41 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,Brent Creech,update,disabled,,true\nThu Sep 08 2022 19:03:16 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,Brent Creech,update,accessLevel,\"{\"\"accessLevel\"\":\"\"manage\"\"}\",\"{\"\"accessLevel\"\":\"\"administrator\"\"}\"\nThu Sep 08 2022 18:07:09 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,flow,Inbound EDI,update,routers.0.routeRecordsUsing,\"\"\"script\"\"\",\"\"\"input_filters\"\"\"\nThu Sep 08 2022 18:01:34 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signin,IP address,,\"\"\"146.168.17.239\"\"\"\nThu Sep 08 2022 16:42:02 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signout,IP address,,\"\"\"146.168.17.239\"\"\"\nThu Sep 08 2022 15:39:57 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signin,IP address,,\"\"\"146.168.17.239\"\"\"\nThu Sep 08 2022 14:01:27 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signout,IP address,,\"\"\"146.168.17.239\"\"\"\nThu Sep 08 2022 14:00:08 GMT+0000 (Coordinated Universal Time),system,Gene Carbone,script,Determine X12 Document Type,update,scriptContentsHash,\"\"\"612b77a84d644e60471df073027d58629bcbb9adb358a4404a588801a8963af3\"\"\",\"\"\"612b77a84d644e60471df073666c09da24b91e6d92375e178b383da033528940\"\"\"\nThu Sep 08 2022 14:00:08 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,flow,Inbound EDI,update,routers,\"[{\"\"script\"\":{\"\"_scriptId\"\":\"\"6319efae9d81d57023954c3c\"\",\"\"function\"\":\"\"branching\"\"},\"\"routeRecordsUsing\"\":\"\"script\"\",\"\"id\"\":\"\"EnNOTE\"\",\"\"routeRecordsTo\"\":\"\"first_matching_branch\"\",\"\"branches\"\":[{\"\"name\"\":\"\"X12 Version 4010\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}]},{\"\"name\"\":\"\"All Other Versions\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}]}]}]\",\"[{\"\"routeRecordsUsing\"\":\"\"script\"\",\"\"id\"\":\"\"EnNOTE\"\",\"\"routeRecordsTo\"\":\"\"first_matching_branch\"\",\"\"branches\"\":[{\"\"name\"\":\"\"X12 Version 4010\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}],\"\"nextRouterId\"\":\"\"i8nIPM\"\"},{\"\"name\"\":\"\"All Other Versions\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}]}],\"\"script\"\":{\"\"_scriptId\"\":\"\"6319efae9d81d57023954c3c\"\",\"\"function\"\":\"\"branching\"\"}},{\"\"routeRecordsUsing\"\":\"\"script\"\",\"\"id\"\":\"\"i8nIPM\"\",\"\"routeRecordsTo\"\":\"\"first_matching_branch\"\",\"\"branches\"\":[{\"\"name\"\":\"\"Purchase Order\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}]},{\"\"name\"\":\"\"Invoice\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}]},{\"\"name\"\":\"\"Advance Ship Notice\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}]},{\"\"name\"\":\"\"Unhandled Document\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}]}],\"\"script\"\":{\"\"_scriptId\"\":\"\"6319f2963437844a8b0f9608\"\",\"\"function\"\":\"\"branching\"\"}}]\"\nThu Sep 08 2022 13:48:06 GMT+0000 (Coordinated Universal Time),system,Gene Carbone,script,Determine X12 Document Type,create,,,\nThu Sep 08 2022 13:41:26 GMT+0000 (Coordinated Universal Time),system,Gene Carbone,script,Determine X12 Version,update,scriptContentsHash,\"\"\"612b77a84d644e60471df073027d58629bcbb9adb358a4404a588801a8963af3\"\"\",\"\"\"612b77a84d644e60471df0739262a430fe4788658411efd7530e89887166c8d9\"\"\"\nThu Sep 08 2022 13:41:20 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,flow,Inbound EDI,update,routers,[],\"[{\"\"routeRecordsUsing\"\":\"\"script\"\",\"\"id\"\":\"\"EnNOTE\"\",\"\"routeRecordsTo\"\":\"\"first_matching_branch\"\",\"\"branches\"\":[{\"\"name\"\":\"\"X12 Version 4010\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}]},{\"\"name\"\":\"\"All Other Versions\"\",\"\"pageProcessors\"\":[{\"\"responseMapping\"\":{\"\"fields\"\":[],\"\"lists\"\":[]},\"\"setupInProgress\"\":true}]}],\"\"script\"\":{\"\"_scriptId\"\":\"\"6319efae9d81d57023954c3c\"\",\"\"function\"\":\"\"branching\"\"}}]\"\nThu Sep 08 2022 13:35:42 GMT+0000 (Coordinated Universal Time),system,Gene Carbone,script,Determine X12 Version,create,,,\nThu Sep 08 2022 13:25:12 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,flow,Inbound EDI,update,pageGenerators,[],\"[{\"\"_exportId\"\":\"\"6319ed36a9a1922e7f1bab81\"\",\"\"skipRetries\"\":false}]\"\nThu Sep 08 2022 13:25:12 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,export,Retrieve Inbound EDI Documents,update,rawData,,\"\"\"612b77a84d644e60471df0732b9bc9c2eabd4a6286e13c213e85e937\"\"\"\nThu Sep 08 2022 13:25:11 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,export,Retrieve Inbound EDI Documents,create,,,\nThu Sep 08 2022 13:23:39 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signin,IP address,,\"\"\"146.168.17.239\"\"\"\nThu Sep 08 2022 13:16:33 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signout,IP address,,\"\"\"146.168.17.239\"\"\"\nThu Sep 08 2022 13:16:19 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signin,IP address,,\"\"\"146.168.17.239\"\"\"\nThu Sep 08 2022 13:14:28 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,integration,Branching Examples,update,_registeredConnectionIds,[],\"[\"\"6319eab44c77df2c045c5127\"\"]\"\nThu Sep 08 2022 13:14:28 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,connection,Orderful Demo,create,,,\nThu Sep 08 2022 13:13:05 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signin,IP address,,\"\"\"146.168.17.239\"\"\"\nThu Sep 08 2022 13:12:50 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signout,IP address,,\"\"\"146.168.17.239\"\"\"\nThu Sep 08 2022 13:10:54 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,connection,GGC celigo.files.com,update,ftp.password,\"\"\"******\"\"\",\"\"\"******\"\"\"\nThu Sep 08 2022 13:09:35 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,flow,Inbound EDI,create,,,\nThu Sep 08 2022 13:09:04 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,integration,Branching Examples,create,,,\nThu Sep 08 2022 13:07:30 GMT+0000 (Coordinated Universal Time),ui,Gene Carbone,user,,signin,IP address,,\"\"\"146.168.17.239\"\"\"\n"
                }
              }
            }
          }
        },
        "tags": []
      },
      "post": {
        "summary": "/filesystem",
        "responses": {},
        "tags": []
      }
    },
...
    "/filesystem/listdir": {
      "get": {
        "summary": "/filesystem/listdir",
        "parameters": [
          {
            "name": "dir",
            "in": "query",
            "example": ".",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filesOnly",
            "in": "query",
            "example": "true",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "path": {
                        "type": "string"
                      },
                      "size": {
                        "type": "number"
                      },
                      "lastModifiedTime": {
                        "type": "string"
                      },
                      "createdTime": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "type",
                      "name",
                      "path",
                      "size",
                      "lastModifiedTime",
                      "createdTime"
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": []
      }
    },
...
  },
...
  "servers": [
    {
      "url": "https://localhost:4000"
    }
  ]
}